npm run build: Enable webpack for easier JS linkage
We want to enable the behavior:

<script src="https://cdn.jsdeliver.net/npm/cardano-dapp-js..."></script>

So that even rudimentary HTML/CSS applications can include the static
webpacked library and integrate a dApp connector.  This change
introduces webpack in order to build into the dist/ directory for each
time we publish.

NOTE: `dist/` is in .gitignore, but "npm publish" should pick this up
and publish to the repository.

[ Documentation: README.md ]
[ Testing: Manual inspection of files output to dist/ ]