src/examples: Add a simple ADA transfer web page example [1.0.4]
To show the functionality of the wallet picker and how to combine it
with Lucid/Blockfrost, we add a new file in `src/examples/` that
includes a simple HTML layout and enables a user to send ADA on submit.

NOTE: We have to re-export Lucid and Blockfrost because we have chosen
not to encapsulate them in the library.  This is a code smell, so as
this project evolves we will have to investigate how to fix.