cardano-dapp-js: Post a message when new successful connection
This change pushes a new message out when there is a successful wallet
connection.  Rather than mess with asynchronous callbacks (which can be
tricky in their expected behavior and timings), we simply allow other
Javascript functions to receive messages from the window.  This is
convenient as well in allowing third-party libraries to listen for
wallet change events and integrate with this dApp connector
automagically.

NOTE: We cannot send the entire wallet object because CIP-30 returns
something that is not serializable.  Instead, we return information
describing the wallet that was picked.

[ Documentation: README.md ]
[ Testing: Manual using `npm link` ]